fx quant: do not insert observers at quantized inputs
Summary:

Context: the existing implementation of `quantized_input_idxs` is convert-only.
Therefore, observers are inserted between the input and the first
quantized node.  This is a problem during QAT, because the initial
input is a fake_quant, and it starts with scale=1 and zp=0.  This does
not match the quantization parameters of the graph input, which can
lead to incorrect numerics.

Fix: do not insert observer for a quantized input.

Test Plan:

```
python test/test_quantization.py TestQuantizeFx
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 936518ba30591e75900c36168183f68f7e97e5ab
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49239