fx quant: make sure observer is inserted before a quantized output
Summary:

Before: if an output was marked as quantized, it could actually not
be quantized, if the previous node was not quantized.

After: if an output was marked as quantized, it will be quantized
regardless of the quantization status of the previous node.

Test Plan:

```
python test/test_quantization.py TestQuantizeFxOps.test_quant_output_always_observed
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 09cf0bc7ea475a0266c2cd03e008976a383a00d5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49420